fabrics: move libnvmf_context to private-fabrics header
Prepare the stage for generating getter/setters for libnvmf_context. For
this, the struct needs to be in the private-fabrics header so the
getter/setters are placed in the correct header file.
