+++ libnvme/src/nvme/private-fabrics.h
@@ -9,6 +9,63 @@
 
 #include <nvme/tree.h>
 
+struct libnvmf_context {
+	struct libnvme_global_ctx *ctx;
+
+	/* common callbacks */
+	bool (*decide_retry)(struct libnvmf_context *fctx, int err,
+			void *user_data);
+	void (*connected)(struct libnvmf_context *fctx, struct libnvme_ctrl *c,
+			void *user_data);
+	void (*already_connected)(struct libnvmf_context *fctx,
+			struct libnvme_host *host, const char *subsysnqn,
+			const char *transport, const char *traddr,
+			const char *trsvcid, void *user_data);
+
+	/* discovery callbacks */
+	void (*discovery_log)(struct libnvmf_context *fctx,
+			bool connect,
+			struct nvmf_discovery_log *log,
+			uint64_t numrec, void *user_data);
+	int (*parser_init)(struct libnvmf_context *fctx,
+			void *user_data);
+	void (*parser_cleanup)(struct libnvmf_context *fctx,
+			void *user_data);
+	int (*parser_next_line)(struct libnvmf_context *fctx,
+			void *user_data);
+
+	/* discovery defaults */
+	int default_max_discovery_retries;
+	int default_keep_alive_timeout;
+
+	/* common fabrics configuraiton */
+	const char *device;
+	bool persistent;
+	struct libnvme_fabrics_config *cfg;
+
+	/* connection configuration */
+	const char *subsysnqn;
+	const char *transport;
+	const char *traddr;
+	const char *trsvcid;
+	const char *host_traddr;
+	const char *host_iface;
+
+	/* host configuration */
+	const char *hostnqn;
+	const char *hostid;
+
+	/* authentication and transport encryption configuration */
+	const char *hostkey;
+	const char *ctrlkey;
+	const char *keyring;
+	char *tls_key;
+	const char *tls_key_identity;
+
+	void *user_data;
+};
+
+
 /**
  * NVMe-oF private struct definitions.
  *
